requirements.txt
biopython
matplotlib
numpy
pandas
requests
scikit-learn
scipy
seaborn
statsmodels
